refactor(frontend): extract shared useSessionCamperPersons hook
Dedup attendee-fetching query logic and ExpandedAttendee interface
from CreateRequestModal and EditableRequestTarget into a shared hook.
Extract dropdown position calculation into useCallback to eliminate
~30 lines of duplicated math. Fixes latent cache conflict where both
components shared a queryKey but returned different types.
