+++ frontend/src/components/EditableRequestTarget.tsx
@@ -1,12 +1,10 @@
 import { useState, useRef, useEffect, useLayoutEffect, useMemo, useCallback } from 'react'
 import { ChevronDown, ChevronUp, Search, User, ExternalLink, Quote } from 'lucide-react'
-import { useQuery } from '@tanstack/react-query'
-import { pb } from '../lib/pocketbase'
-import type { PersonsResponse, AttendeesResponse } from '../types/pocketbase-types'
+import type { PersonsResponse } from '../types/pocketbase-types'
 import { getDisplayAgeForYear } from '../utils/displayAge'
 import clsx from 'clsx'
 import { useClickOutside } from '../hooks/useClickOutside'
-import { queryKeys } from '../utils/queryKeys'
+import { useSessionCamperPersons } from '../hooks/useSessionCamperPersons'
 
 interface EditableRequestTargetProps {
   requestType: string
@@ -79,109 +77,71 @@ export default function EditableRequestTarget({
   const searchInputRef = useRef<HTMLInputElement>(null)
 
   // Calculate dropdown position on open - use fixed positioning to escape overflow containers
-  useLayoutEffect(() => {
-    if (isOpen && triggerRef.current) {
-      const triggerRect = triggerRef.current.getBoundingClientRect()
-      const viewportHeight = window.innerHeight
-      const dropdownHeight = 400 // Approximate max height of dropdown
+  // Calculate dropdown position relative to trigger — shared by initial open and scroll/resize
+  const calculatePosition = useCallback(() => {
+    if (!triggerRef.current) return
+    const triggerRect = triggerRef.current.getBoundingClientRect()
+    const viewportHeight = window.innerHeight
+    const dropdownHeight = 400 // Approximate max height of dropdown
 
-      const spaceBelow = viewportHeight - triggerRect.bottom
-      const spaceAbove = triggerRect.top
+    const spaceBelow = viewportHeight - triggerRect.bottom
+    const spaceAbove = triggerRect.top
 
-      // Open upward if not enough space below and more space above
-      if (spaceBelow < dropdownHeight && spaceAbove > spaceBelow) {
-        setDropdownPosition({
-          bottom: viewportHeight - triggerRect.top + 4,
-          left: triggerRect.left,
-          direction: 'up',
-        })
-      } else {
-        setDropdownPosition({
-          top: triggerRect.bottom + 4,
-          left: triggerRect.left,
-          direction: 'down',
-        })
-      }
+    if (spaceBelow < dropdownHeight && spaceAbove > spaceBelow) {
+      setDropdownPosition({
+        bottom: viewportHeight - triggerRect.top + 4,
+        left: triggerRect.left,
+        direction: 'up',
+      })
+    } else {
+      setDropdownPosition({
+        top: triggerRect.bottom + 4,
+        left: triggerRect.left,
+        direction: 'down',
+      })
     }
-  }, [isOpen])
+  }, [])
 
-  // Recalculate position on scroll (parent container may scroll)
+  // Position on open
+  useLayoutEffect(() => {
+    if (isOpen) calculatePosition()
+  }, [isOpen, calculatePosition])
+
+  // Recalculate position on scroll (parent container may scroll) and resize
   useEffect(() => {
     if (!isOpen) return
 
-    const recalculatePosition = () => {
-      if (triggerRef.current) {
-        const triggerRect = triggerRef.current.getBoundingClientRect()
-        const viewportHeight = window.innerHeight
-        const dropdownHeight = 400
-
-        const spaceBelow = viewportHeight - triggerRect.bottom
-        const spaceAbove = triggerRect.top
-
-        if (spaceBelow < dropdownHeight && spaceAbove > spaceBelow) {
-          setDropdownPosition({
-            bottom: viewportHeight - triggerRect.top + 4,
-            left: triggerRect.left,
-            direction: 'up',
-          })
-        } else {
-          setDropdownPosition({
-            top: triggerRect.bottom + 4,
-            left: triggerRect.left,
-            direction: 'down',
-          })
-        }
-      }
-    }
-
-    // Listen for scroll on any ancestor
-    window.addEventListener('scroll', recalculatePosition, true)
-    window.addEventListener('resize', recalculatePosition)
+    window.addEventListener('scroll', calculatePosition, true)
+    window.addEventListener('resize', calculatePosition)
 
     return () => {
-      window.removeEventListener('scroll', recalculatePosition, true)
-      window.removeEventListener('resize', recalculatePosition)
+      window.removeEventListener('scroll', calculatePosition, true)
+      window.removeEventListener('resize', calculatePosition)
     }
-  }, [isOpen])
+  }, [isOpen, calculatePosition])
 
   // Fetch session campers for person selection
-  const { data: allCampers = [] } = useQuery({
-    queryKey: queryKeys.sessionCampers(sessionId, year),
-    queryFn: async () => {
-      // Filter attendees by session_id (CampMinder ID) server-side
-      const attendees = await pb.collection<AttendeesResponse>('attendees').getFullList({
-        filter: `session_id = ${sessionId} && year = ${year} && status = "enrolled"`,
-        expand: 'person',
-      })
-
-      interface ExpandedAttendee {
-        person?: PersonsResponse
-      }
-
-      // Map to camper format with session info
-      return attendees
-        .map((attendee) => {
-          const expanded = attendee.expand as ExpandedAttendee | undefined
-          const person = expanded?.person
-          if (!person) return null
-          const camper: Camper = {
-            id: person.id,
-            campminder_person_id: person.cm_id,
-            first_name: person.first_name,
-            last_name: person.last_name,
-            preferred_name: person.preferred_name,
-            age: person.age,
-            birthdate: person.birthdate,
-            grade: person.grade ?? 0,
-            gender: person.gender || '',
-            session_cm_id: sessionId,
-          }
-          return camper
-        })
-        .filter(Boolean) as Camper[]
-    },
+  const { data: camperPersons = [] } = useSessionCamperPersons(sessionId, year, {
     enabled: requestType !== 'age_preference' && isOpen,
   })
+
+  // Map PersonsResponse to local Camper shape
+  const allCampers: Camper[] = useMemo(
+    () =>
+      camperPersons.map((person) => ({
+        id: person.id,
+        campminder_person_id: person.cm_id,
+        first_name: person.first_name,
+        last_name: person.last_name,
+        preferred_name: person.preferred_name,
+        age: person.age,
+        birthdate: person.birthdate,
+        grade: person.grade ?? 0,
+        gender: person.gender || '',
+        session_cm_id: sessionId,
+      })),
+    [camperPersons, sessionId]
+  )
 
   // Get current person from personMap (passed from parent) instead of separate query
   const currentPerson = useMemo(() => {
